requirements.txt
pyserial
pynmea2
python-socketio==4.6.1      # last version to support Socket.IO v2 (https://pypi.org/project/python-socketio/)
configparser
pytest
python-dotenv